requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blessed=1.20.0=pypi_0
brotli=1.1.0=pypi_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2023.7.22=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
cmake=3.25.0=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.10.0=pypi_0
dataclasses-json=0.6.6=pypi_0
dds-cloudapi-sdk=0.2.1=pypi_0
defusedxml=0.7.1=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
easydict=1.13=pypi_0
einops=0.8.0=pypi_0
email-validator=2.1.1=pypi_0
exceptiongroup=1.2.1=pypi_0
fairscale=0.4.4=pypi_0
faiss-cpu=1.8.0=pypi_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.51.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
g4f=0.3.1.9=pypi_0
gdino=1.5=dev_0
gpustat=1.1.1=pypi_0
gradio=4.22.0=pypi_0
gradio-client=0.13.0=pypi_0
groundingdino=0.1.0=dev_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.0=pypi_0
idna=2.10=pypi_0
imageio=2.34.1=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
jinja2=3.1.3=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lit=15.0.7=pypi_0
litellm=1.37.17=pypi_0
loguru=0.7.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.2=pypi_0
matplotlib=3.8.4=pypi_0
mdurl=0.1.2=pypi_0
mmcls=0.25.0=pypi_0
mmcv-full=1.5.0=pypi_0
mmsegmentation=0.27.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-ml-py=12.550.52=pypi_0
omegaconf=2.3.0=pypi_0
onemetric=0.1.2=pypi_0
openai=1.30.1=pypi_0
opencv-contrib-python=4.9.0.80=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.8.0.74=pypi_0
openssl=3.0.13=h7f8727e_1
orjson=3.10.3=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.2.0=pypi_0
pip=24.0=py39h06a4308_0
platformdirs=4.2.1=pypi_0
prettytable=3.10.0=pypi_0
psutil=5.9.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.7=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.6.3=pypi_0
pydantic-core=2.16.3=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-magic=0.4.27=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
ram=0.0.1=dev_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.10=pypi_0
requests=2.31.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.7.1=pypi_0
roboflow=1.1.28=pypi_0
rpds-py=0.18.1=pypi_0
ruff=0.4.4=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.22.0=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
segment-anything=1.0=pypi_0
segment-anything-hq=0.3=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=69.5.1=py39h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.37.2=pypi_0
supervision=0.20.0=pypi_0
sympy=1.12=pypi_0
tifffile=2024.5.22=pypi_0
tiktoken=0.7.0=pypi_0
timm=0.4.12=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.0.0+cu118=pypi_0
torchaudio=2.0.1+cu118=pypi_0
torchvision=0.15.1+cu118=pypi_0
tqdm=4.66.4=pypi_0
transformers=4.25.1=pypi_0
trex=1.0=dev_0
triton=2.0.0=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.9.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
ujson=5.10.0=pypi_0
ultralytics=8.0.120=pypi_0
urllib3=1.26.13=pypi_0
uvicorn=0.29.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.21.0=pypi_0
wcwidth=0.2.13=pypi_0
websockets=11.0.3=pypi_0
wheel=0.43.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
yapf=0.40.2=pypi_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_1
